Reorganize crashlytics utils
