+++ app/src/fdroid/java/io/github/wulkanowy/utils/CrashlyticsUtils.kt
@@ -1,5 +1,8 @@
-package io.github.wulkanowy.utils.logger
+@file:Suppress("UNUSED_PARAMETER")
 
+package io.github.wulkanowy.utils
+
+import android.content.Context
 import timber.log.Timber
 
 class CrashlyticsTree : Timber.Tree() {
@@ -7,4 +10,8 @@ class CrashlyticsTree : Timber.Tree() {
     override fun log(priority: Int, tag: String?, message: String, t: Throwable?) {
         // do nothing
     }
+}
+
+fun initCrashlytics(context: Context) {
+    // do nothing
 }
