--- app/src/fdroid/java/io/github/wulkanowy/utils/logger/LoggerHelper.kt
@@ -1,8 +1,0 @@
-package io.github.wulkanowy.utils.logger
-
-import android.content.Context
-
-@Suppress("UNUSED_PARAMETER")
-fun initCrashlytics(context: Context) {
-    // do nothing
-}
