+++ app/src/main/java/io/github/wulkanowy/di/AppModule.kt
@@ -6,7 +6,7 @@ import dagger.Module
 import dagger.Provides
 import eu.davidea.flexibleadapter.FlexibleAdapter
 import eu.davidea.flexibleadapter.items.AbstractFlexibleItem
-import io.github.wulkanowy.BuildConfig
+import io.github.wulkanowy.BuildConfig.DEBUG
 import io.github.wulkanowy.WulkanowyApp
 import io.github.wulkanowy.utils.SchedulersProvider
 import javax.inject.Named
@@ -14,6 +14,7 @@ import javax.inject.Singleton
 
 @Module
 internal class AppModule {
+
     @Singleton
     @Provides
     fun provideContext(app: WulkanowyApp): Context = app
@@ -32,5 +33,5 @@ internal class AppModule {
     @Singleton
     @Named("isDebug")
     @Provides
-    fun provideIsDebug() = BuildConfig.DEBUG
+    fun provideIsDebug() = DEBUG
 }
