--- app/src/main/java/io/github/wulkanowy/utils/logger/DebugLogTree.kt
@@ -1,10 +1,0 @@
-package io.github.wulkanowy.utils.logger
-
-import timber.log.Timber
-
-class DebugLogTree : Timber.DebugTree() {
-
-    override fun log(priority: Int, tag: String?, message: String, t: Throwable?) {
-        super.log(priority, "Wulkanowy", message, t)
-    }
-}
